+++ src/App.js
@@ -1,31 +1,16 @@
 // import * as css from "./scss/index.module.scss";
 import "./scss/index.scss";
-import UserName from "./images/user.png";
-
-const user = {
-  name: "Hedy Lamarr",
-  imageUrl: UserName,
-  imageSize: 90,
-};
-
-const MyButton = () => {
-  return <button className="button">I'm a button</button>;
-};
+// import UserName from "./images/user.png";
+import Records from "./data.json";
 
 const App = () => {
   return (
     <div className="container">
-      <h1>{user.name}</h1>
-      <img
-        className="avatar"
-        src={user.imageUrl}
-        alt={"Photo of " + user.name}
-        style={{
-          width: user.imageSize,
-          height: user.imageSize,
-        }}
-      />
-      <MyButton />
+      <ul>
+        {Records.map((product) => (
+          <li key={product.id}>{product.title}</li>
+        ))}
+      </ul>
     </div>
   );
 };
