+++ src/data.json
@@ -1,0 +1,14 @@
+[
+    { 
+        "title": "Alex", 
+        "id": 1 
+    },
+    { 
+        "title": "Nicos", 
+        "id": 2 
+    },
+    { 
+        "title": "Rafael", 
+        "id": 3 
+    }
+]
